+++ src/pages/Introduce/components/Swipe/style.ts
