+++ src/pages/Main/components/PostList/style.ts
