+++ src/pages/PostCreate/index.tsx
@@ -1,8 +1,8 @@
 import { useParams } from 'react-router-dom';
 import { useEffect } from 'react';
-import PostForm from '@/pages/PostUpdate/component/PostForm';
+import PostForm from '@/pages/PostUpdate/components/PostForm';
 import { FormPageContainer } from '@/pages/PostCreate/style';
-import UserInfo from '@/pages/PostUpdate/component/UserInfo';
+import UserInfo from '@/pages/PostUpdate/components/UserInfo';
 import { getMyInfo } from '@/slices/user/thunk';
 import { useSelectedMyInfo } from '@/hooks/useSelectedMyInfo';
 import { useDispatch } from '@/store';
