+++ src/pages/PostDetail/components/Result/style.ts
