+++ src/pages/PostUpdate/components/VoteBox/style.ts
