+++ src/pages/PostUpdate/index.tsx
@@ -1,10 +1,10 @@
 import { useParams } from 'react-router-dom';
 import { useEffect } from 'react';
-import UpdateForm from '@/pages/PostUpdate/component/UpdateForm';
+import UpdateForm from '@/pages/PostUpdate/components/UpdateForm';
 import { useDispatch } from '@/store';
 import { getPostDetail } from '@/slices/postDetail/thunk';
 import { getMyInfo } from '@/slices/user/thunk';
-import UserInfo from '@/pages/PostUpdate/component/UserInfo';
+import UserInfo from '@/pages/PostUpdate/components/UserInfo';
 import { useSelectedMyInfo } from '@/hooks/useSelectedMyInfo';
 import { FormPageContainer } from '@/pages/PostCreate/style';
