+++ src/pages/Profile/components/GrassTable/style.ts
@@ -1,6 +1,6 @@
+import { LightnessType } from '.';
 import styled from 'styled-components';
 import theme from '@/styles/theme';
-import { LightnessType } from '@/pages/Profile/components/GrassTable/type';
 
 export const GrassWrapper = styled.div`
   width: 152px;
